Not all files are logs
