+++ tools/lib/logreader.py
@@ -18,9 +18,8 @@ from urllib.parse import parse_qs, urlparse
 
 from cereal import log as capnp_log
 from openpilot.common.swaglog import cloudlog
-from openpilot.tools.lib.comma_car_segments import get_url as get_comma_segments_url
-from openpilot.tools.lib.openpilotci import get_url
-from openpilot.tools.lib.filereader import DATA_ENDPOINT, FileReader, file_exists, internal_source_available
+from openpilot.tools.lib.filereader import DATA_ENDPOINT, FileReader, file_exists, internal_source_available, FilePath
+from openpilot.tools.lib.file_sources import comma_api_source
 from openpilot.tools.lib.route import Route, SegmentRange, FileName
 from openpilot.tools.lib.log_time_series import msgs_to_time_series
 
@@ -86,7 +85,6 @@ class CachedEventReader:
       return getattr(self, name)
     return getattr(self._evt, name)
 
-
 class _LogFileReader:
   def __init__(self, fn, only_union_types=False, sort_by_time=False, dat=None):
     self.data_version = None
@@ -131,76 +129,16 @@ class _LogFileReader:
       else:
         yield ent
 
-
 class ReadMode(enum.StrEnum):
   RLOG = "r"  # only read rlogs
   QLOG = "q"  # only read qlogs
   AUTO = "a"  # default to rlogs, fallback to qlogs
   AUTO_INTERACTIVE = "i"  # default to rlogs, fallback to qlogs with a prompt from the user
 
-
-LogPath = str | None
-LogFileName = tuple[str, ...]
-Source = Callable[[SegmentRange, list[int], LogFileName], dict[int, LogPath]]
-
 InternalUnavailableException = Exception("Internal source not available")
 
 
-class LogsUnavailable(Exception):
-  pass
-
-
-def comma_api_source(sr: SegmentRange, seg_idxs: list[int], fns: LogFileName) -> dict[int, LogPath]:
-  route = Route(sr.route_name)
-
-  # comma api will have already checked if the file exists
-  if fns == FileName.RLOG:
-    return {seg: route.log_paths()[seg] for seg in seg_idxs}
-  else:
-    return {seg: route.qlog_paths()[seg] for seg in seg_idxs}
-
-
-def internal_source(sr: SegmentRange, seg_idxs: list[int], fns: LogFileName, endpoint_url: str = DATA_ENDPOINT) -> dict[int, LogPath]:
-  if not internal_source_available(endpoint_url):
-    raise InternalUnavailableException
-
-  def get_internal_url(sr: SegmentRange, seg, file):
-    return f"{endpoint_url.rstrip('/')}/{sr.dongle_id}/{sr.log_id}/{seg}/{file}"
-
-  return eval_source({seg: [get_internal_url(sr, seg, fn) for fn in fns] for seg in seg_idxs})
-
-
-def openpilotci_source(sr: SegmentRange, seg_idxs: list[int], fns: LogFileName) -> dict[int, LogPath]:
-  return eval_source({seg: [get_url(sr.route_name, seg, fn) for fn in fns] for seg in seg_idxs})
-
-
-def comma_car_segments_source(sr: SegmentRange, seg_idxs: list[int], fns: LogFileName) -> dict[int, LogPath]:
-  return eval_source({seg: get_comma_segments_url(sr.route_name, seg) for seg in seg_idxs})
-
-
-def direct_source(file_or_url: str) -> list[str]:
-  return [file_or_url]
-
-
-def eval_source(files: dict[int, list[str] | str]) -> dict[int, LogPath]:
-  # Returns valid file URLs given a list of possible file URLs for each segment (e.g. rlog.bz2, rlog.zst)
-  valid_files: dict[int, LogPath] = {}
-
-  for seg_idx, urls in files.items():
-    if isinstance(urls, str):
-      urls = [urls]
-
-    # Add first valid file URL or None
-    for url in urls:
-      if file_exists(url):
-        valid_files[seg_idx] = url
-        break
-    else:
-      valid_files[seg_idx] = None
-
-  return valid_files
-
-
+# TODO this should apply to camera files as well
 def auto_source(identifier: str, sources: list[Source], default_mode: ReadMode) -> list[str]:
   exceptions = {}
